# contracts/feedback/openapi.yaml
openapi: 3.1.0
...
info:
  title: Feedback Service API
  version: 1.0.0-alpha.1
  description: |
    ホスト業務システムから独立したフィードバック基盤の公開契約。
    Web GIS の project、GIS、業務 CRUD、タイル API を含めない。
servers:
  - url: /feedback/v1
security:
  - bearerAuth: []
...
paths:
  /capabilities:
    get:
      tags: [system]
      operationId: getFeedbackCapabilities
      security: []
      responses:
        "200":
          description: API と schema の対応能力
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackCapabilities"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /me:
    get:
      tags: [identity]
      operationId: getFeedbackMe
      responses:
        "200":
          description: 認証主体と workspace 権限
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackMe"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /applications/{applicationKey}/manifest:
    parameters:
      - $ref: "#/components/parameters/ApplicationKey"
    get:
      tags: [applications]
      operationId: getFeedbackApplicationManifest
      responses:
        "200":
          description: 登録済み application manifest
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackApplicationManifestV1"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [applications]
      operationId: putFeedbackApplicationManifest
      parameters:
        - $ref: "#/components/parameters/IfMatchOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackApplicationManifestV1"
      responses:
        "200":
          description: 登録した manifest
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackApplicationManifestV1"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /review-context:
    get:
      tags: [review-context]
      operationId: getFeedbackReviewContext
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/EnvironmentKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
        - name: release
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
        - name: locale
          in: query
          schema: { type: string, maxLength: 35 }
        - name: pageKey
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
        - name: routeTemplate
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 500 }
        - name: pathParameters
          in: query
          required: true
          description: JSON object として直列化した path parameter。
          schema: { type: string, minLength: 2, maxLength: 4000 }
        - name: queryParameters
          in: query
          description: JSON object として直列化した許可候補 query parameter。
          schema: { type: string, maxLength: 8000 }
      responses:
        "200":
          description: 現在位置へ適用するレビュー文脈
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReviewContextV1"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        直接OIDC JWTまたは契約済みtoken exchange JWT。直接OIDC JWTは検証対象のissuer、audience、
        時刻、設定済みsubject claimに加え、`feedback_permissions`を1件以上のFeedbackPermissionからなる
        文字列配列として必須とする。直接OIDCのpermissionは全resourceでDB membershipと交差する。
        token exchange JWTのresource scopeとclaimはtoken exchange公開契約に従う。
  parameters:
    ApplicationKey:
      name: applicationKey
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z][a-z0-9-]{0,62}$" }
    ApplicationKeyQuery:
      name: applicationKey
      in: query
      required: true
      schema: { type: string, pattern: "^[a-z][a-z0-9-]{0,62}$" }
    EnvironmentKeyQuery:
      name: environmentKey
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
    ExternalWorkspaceKeyQuery:
      name: externalWorkspaceKey
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
...
    IfMatchOptional:
      name: If-Match
      in: header
      required: false
      schema: { type: string, minLength: 3, maxLength: 200 }
...
  headers:
    ETag:
      description: resource version。更新時は If-Match へそのまま指定する。
      schema: { type: string }
  responses:
    BadRequest:
      description: 契約または manifest に適合しない入力
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Unauthorized:
      description: 未認証または token 無効
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Forbidden:
      description: feedback permission 不足
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    NotFound:
      description: resource 不在または存在秘匿
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
    PreconditionFailed:
      description: ETag が古い
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
    ServiceUnavailable:
      description: DB または必須依存が利用不能
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
  schemas:
    FeedbackApplicationManifestV1:
      $ref: "./schemas/application-manifest.schema.json"
...
    FeedbackCapabilities:
      type: object
      additionalProperties: false
      required: [apiVersion, apiMajorVersion, manifestSchemaVersions, targetSchemaVersions, evidence, features]
      properties:
        apiVersion: { type: string, const: "1.0" }
        apiMajorVersion: { type: integer, const: 1 }
        manifestSchemaVersions:
          type: array
          items: { type: string, const: "1" }
        targetSchemaVersions:
          type: array
          items: { type: string, const: "1" }
        evidence:
          type: object
          additionalProperties: false
          required: [maxBytes, maxCountPerWorkspace, acceptedContentTypes]
          properties:
            maxBytes: { type: integer, minimum: 1 }
            maxCountPerWorkspace: { type: integer, minimum: 1 }
            acceptedContentTypes:
              type: array
              items: { enum: [image/png, image/webp] }
        features:
          type: array
          items: { type: string }
    FeedbackProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string, pattern: "^[a-z0-9_.-]+$" }
        requestId: { type: string, minLength: 1, maxLength: 200 }
    FeedbackPermission:
      enum: [feedback.read, feedback.comment, feedback.manage, feedback.admin]
    FeedbackParticipant:
      type: object
      additionalProperties: false
      required: [principalId]
      properties:
        principalId: { type: string, minLength: 1, maxLength: 200 }
        displayName: { type: [string, "null"], maxLength: 200 }
        participantName: { type: [string, "null"], maxLength: 100 }
...
    FeedbackMembership:
      type: object
      additionalProperties: false
      required: [applicationKey, externalWorkspaceKey, permissions]
      properties:
        applicationKey: { type: string }
        externalWorkspaceKey: { type: string }
        permissions:
          type: array
          items: { $ref: "#/components/schemas/FeedbackPermission" }
    FeedbackMe:
      type: object
      additionalProperties: false
      required: [participant, memberships]
      properties:
        participant: { $ref: "#/components/schemas/FeedbackParticipant" }
        memberships:
          type: array
          items: { $ref: "#/components/schemas/FeedbackMembership" }
...
    FeedbackParticipantPolicy:
      oneOf:
        - type: object
          additionalProperties: false
          required: [mode]
          properties:
            mode: { const: authenticated-identity }
        - type: object
          additionalProperties: false
          required: [mode, storage]
          properties:
            mode: { enum: [prompt, authenticated-and-prompt] }
            storage: { enum: [memory, local-storage, host] }
    FeedbackEvidencePolicy:
      type: object
      additionalProperties: false
      required: [enabled, maxBytes, acceptedContentTypes]
      properties:
        enabled: { type: boolean }
        maxBytes: { type: integer, minimum: 1 }
        acceptedContentTypes:
          type: array
          items: { enum: [image/png, image/webp] }
    FeedbackReviewContextV1:
      type: object
      additionalProperties: false
      required: [session, scope, posting, permissions, participantPolicy, evidencePolicy]
      properties:
        session:
          oneOf:
            - $ref: "#/components/schemas/FeedbackSessionV1"
            - type: "null"
        scope: { enum: [reviewable, excluded, unregistered] }
        posting: { enum: [allow, warn, deny] }
        permissions:
          type: array
          items: { $ref: "#/components/schemas/FeedbackPermission" }
        participantPolicy: { $ref: "#/components/schemas/FeedbackParticipantPolicy" }
        evidencePolicy: { $ref: "#/components/schemas/FeedbackEvidencePolicy" }
    FeedbackSessionStatus:
      type: string
      enum: [draft, open, closed]
    FeedbackSessionV1:
      type: object
      additionalProperties: false
      required:
        [id, applicationKey, environmentKey, externalWorkspaceKey, manifestVersion, title, status,
         outOfScopePosting, scopes, perspectives, createdAt, updatedAt, version]
      properties:
        id: { type: string, format: uuid }
        applicationKey: { type: string }
        environmentKey: { type: string }
        externalWorkspaceKey: { type: string }
        manifestVersion: { type: string }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 5000 }
        status: { $ref: "#/components/schemas/FeedbackSessionStatus" }
        outOfScopePosting: { enum: [allow, warn, deny] }
        startAt: { type: [string, "null"], format: date-time }
        endAt: { type: [string, "null"], format: date-time }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionScopeV1" }
        perspectives:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionPerspectiveV1" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
    FeedbackSessionScopeV1:
      type: object
      additionalProperties: false
      required: [pageKey, reviewable]
      properties:
        pageKey: { type: string, minLength: 1, maxLength: 100 }
        routeTemplate: { type: [string, "null"], maxLength: 500 }
        reviewable: { type: boolean }
        perspectiveCodes:
          type: array
          uniqueItems: true
          maxItems: 100
          description: この画面で確認する観点コード。空配列または省略時は有効な観点すべて。
          items: { type: string, minLength: 1, maxLength: 100 }
    FeedbackSessionPerspectiveV1:
      type: object
      additionalProperties: false
      required: [code, label, status]
      properties:
        code: { type: string, minLength: 1, maxLength: 100 }
        label: { type: string, minLength: 1, maxLength: 200 }
        status: { enum: [active, future, out-of-scope] }
        guidance: { type: [string, "null"], maxLength: 5000 }
...